fix(gateway): a live screen read always asks the owning Director
Inspection 01, finding 1. The stored screen could answer 'what is on screen
right now?' while three facts held, and the first of them - the pushed
TotalBufferBytes still equalling the mark taken at capture - is refreshed by a
ten-second snapshot timer and by some activity transitions, never by the
terminal being written to. So all three could pass on a screen the terminal had
already scrolled past, and the shipped negative-control test asserted exactly
that stale serve rather than forbidding it.

Resolution 2 of ruling 12: ReadLiveAsync always goes to the tunnel, ReadStored
keeps serving history. The store could only ever answer while the owning
Director's tunnel was CONNECTED, which is exactly when the tunnel could have
answered itself, so the live half never bought availability - only latency on a
connection that was already up.

The negative control is rewritten to forbid the stale serve: a stored screen, a
connected Director, a one-second-old snapshot and equal byte marks, and the
reader must still return the tunnel's rows by content with one tunnel call.
Against the unfixed reader it failed 'Expected: Tunnel / Actual: Store'.
Inspection 01's cross-Director repro ships as a test in its own right.

Row 7 is withdrawn (ruling 13): a voice turn costing no tunnel screen read is
false by design now. Row 6 is replaced by the finding-1 test. The ScreenGridPulls
counter is kept with its purpose inverted in its own comment - it now guards
against a live read being answered from the store again.
